Remove obsolete realsense_navigation package.
The realsense_navigation package contained within this
package is now obsolete, as support for the R200
camera has been added to the native turtlebot_navigation
package.

Please see the Turtlebot wiki (http://wiki.ros.org/Robots/TurtleBot)
for instructions on configuring the Turtlebot to
work with the R200 camera. The default turtlebot_navigation
packages and tutorials should now work with R200.

The r200_nodelet_navigation.launch file was introduced as a
temporary patch while an updated turtlebot_bringup
change was waiting to be merged. It is now obsolete
and has also been removed.