Change default setting for RealSense emitters
The IR emitters for a RealSense camera was set to off by default.
For some reason, this default setting was not working, and in order
to turn the emitters off we had to toggle the emitters twice. By
keeping them on by default, we can turn them off with only one toggle
which is done in `chippy.rosinstall` for Frybot. The Grillbot needs
them to be on.